% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_into_node.R
\name{add.into.node}
\alias{add.into.node}
\title{Function to add tips at random from a given node}
\usage{
add.into.node(tree, node, new.tip, exception.list = NULL, prob = T)
}
\description{
Function to add tips at random from a given node
}
\examples{
exception.list<-list(c("Abies_alba","Abies_pinsapo"),
                     c("Achillea_pyrenaica","Achillea_santolinoides",
                       "Achillea_chamaemelifolia"))
}
