% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_over_node.R
\name{add.over.node}
\alias{add.over.node}
\title{Function to add tips over a node. Valid for "sibling" species or genera.}
\usage{
add.over.node(tree, new.tip, node)
}
\description{
Function to add tips over a node. Valid for "sibling" species or genera.
}
