% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtip_add_random.R
\name{add.random2}
\alias{add.random2}
\title{modification from phytools "add.random". if prob=TRUE, it remains the same;
else a random node is selected without probability vector}
\usage{
add.random2(
  tree,
  n = NULL,
  tips = NULL,
  edge.length = NULL,
  order = "random",
  prob = TRUE
)
}
\arguments{
\item{tips}{Character vector with the labels of the tip to add to the
phylogenetic tree.}

\item{order}{The addition order the tips. If \dQuote{random}, tips are added
at random. If \dQuote{input}, tips are added sequentially.}
}
\description{
modification from phytools "add.random". if prob=TRUE, it remains the same;
else a random node is selected without probability vector
}
