% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_monophyletic.R
\name{add.to.monophyletic}
\alias{add.to.monophyletic}
\title{Function to add tips at random monophyletic clade}
\usage{
add.to.monophyletic(tree, new.tip)
}
\description{
Function to add tips at random monophyletic clade
}
\examples{
#plot(add_to_monophyletic(Tree, "Invent_inventata"))
}
