% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_paraphyletic.R
\name{add.to.paraphyletic}
\alias{add.to.paraphyletic}
\title{Insert in paraphyletic. i no nodes are given, they will be automatically searched.}
\usage{
add.to.paraphyletic(
  tree,
  new.tip,
  prob = TRUE,
  intern.node = NULL,
  group.node = NULL
)
}
\arguments{
\item{intern.node}{mrca node of the clade where not to insert the new tip. It must be used in conjuction
with group node}

\item{group.node}{mrca node of the paraphyletic clade. It must be used in conjuction
with intern node}
}
\description{
Insert in paraphyletic. i no nodes are given, they will be automatically searched.
}
\examples{
#add.to.paraphyletic(Tree, "Achillea_3_inventatus")
}
