% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_singleton.R
\name{add.to.singleton}
\alias{add.to.singleton}
\title{Function to add tip to singleton}
\usage{
add.to.singleton(
  tree,
  singleton,
  new.tips,
  use.singleton = F,
  respect.mono = F,
  respect.para = F
)
}
\description{
Function to add tip to singleton
}
