% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibians.info}
\alias{amphibians.info}
\title{amphibians.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 7238 species of amphibians
included in 'amphibians' object. Taxonomic information was retrieved from 'ncbi' database taxonomy.

}
}
\usage{
amphibians.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the amphibian phylogeny (Pyron et al. 2013)
}
\keyword{datasets}
