% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibians}
\alias{amphibians}
\title{amphibians}
\format{
A phylo object
\describe{
Phylogenetic tree comprising 7238 species of amphibians
Data obtained from https://github.com/megatrees/amphibian_20221117
}
}
\usage{
amphibians
}
\description{
Phylogenetic tree obtained from Jetz & Pyron, (2018).
}
\keyword{datasets}
