% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds.info}
\alias{birds.info}
\title{birds.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 9993 species of birds included in
'birds' object. Taxonomic information was retrieved from 'ncbi' database taxonomy.

}
}
\usage{
birds.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the bird phylogeny (Jetz et al. 2012)
}
\keyword{datasets}
