% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{birds}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 655 species of birds.
Data obtained from https://doi.org/10.3390/d11070109
}
}
\usage{
birds
}
\description{
Phylogenetic tree obtained from Kimball et al. (2019).
}
\keyword{datasets}
