% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{build.info}
\alias{build.info}
\title{Function to create info given a species vector(column)}
\usage{
build.info(
  species,
  tree,
  find.ranks = TRUE,
  db = "ncbi",
  mode = "backbone",
  genus = F
)
}
\description{
Function to create info given a species vector(column)
}
