% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{build.Info1}
\alias{build.Info1}
\title{Function to create DF1 given a species vector(column)}
\usage{
build.Info1(
  species,
  tree,
  find.ranks = TRUE,
  db = "ncbi",
  mode = "backbone",
  genus = F
)
}
\description{
Function to create DF1 given a species vector(column)
}
