% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{build_info}
\alias{build_info}
\title{Create a 'info' data frame.}
\usage{
my.info <- build_info(species = species.list, tree = tree, db="gbif",
                   mode="list", find.ranks = TRUE, interactive =FALSE,
                   genus = FALSE, prior.info = NULL, verbose = TRUE)
}
\arguments{
\item{species}{Character vector or a single-column data frame including
the species of interest. Word breakers must be blanks (" ")
or underscores ("_").}

\item{tree}{A 'phylo' object with the backbone tree (set to NULL if
\code{mode} is "list").}

\item{find.ranks}{Logical. If TRUE, taxonomic information will be retrieved
from the specified taxonomic repository to identify
supra-generic MDCCs for the PUTs.}

\item{db}{Taxonomic repository to query if \code{find.ranks} is
set to TRUE. One of 'ncbi' (default), 'itis', 'gbif'
or 'bold'.}

\item{mode}{If mode is "list", the info data frame will be filled with
the species provided in the \code{species} argument. If mode
is "backbone", the info data frame will also include all the
tips in the backbone tree.}

\item{interactive}{Logical. Whether or not ambiguous species names will
be resolved manually as they appear when retrieving
taxonomic information. If FALSE, NAs will be returned
in the corresponding row of info.}

\item{genus}{Logical. Whether or not the backbone tree is resolved to the
genus level. If TRUE, all the taxa in the tips of the phylogeny
and the species vector must represent genera.}

\item{prior.info}{A previously created info data frame that is recycled to
build the new info.}

\item{verbose}{Logical. Whether or not to print the progress.}
}
\value{
An 'info' data frame
}
\description{
This function creates an 'info' object for a given list of species.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Create a list of species to include in the phylogeny
 catspecies <- c("Lynx_lynx",
"Panthera_uncia",
"Panthera_onca",
"Felis_catus",
"Puma_concolor",
"Lynx_canadensis",
"Panthera_tigris",
"Panthera_leo",
"Felis_silvestris")

#Create the 'info' data frame
cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
