% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cats}
\alias{cats}
\title{cats phylogenetic tree}
\format{
A phylogenetic tree with 14 tips and 13 internal nodes
\describe{
Phylogenetic tree regarding 14 species of felids (members of the Felidae family).
This tree has been extracted from a complete mammal tree by Upham et al. (2019) for example purposes
}
}
\usage{
cats
}
\description{
Phylogenetic tree used for randtip package examples
}
\keyword{datasets}
