% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{check_info}
\alias{check_info}
\title{Function to check randtip's files.}
\usage{
my.check <- check_info(my.info, tree)
}
\arguments{
\item{info}{An 'info' object.}

\item{tree}{The original backbone tree.}

\item{sim}{Name similarity threshold to detect possible misspellings
on tip labels. Default value is 0.85. Similarity is obtained
with \code{stringsim} function from \code{stringdist} package.
See \link[stringdist]{stringsim} for details.}

\item{find.phyleticity}{Logical. Should or not the phyletic nature o the
taxonomic ranks be evaluated.}

\item{search.typos}{Logical. Should or not the function search for possible
misspelling on tip labels. This match lookup will be performed for
all PUTs using all tree tips.}

\item{verbose}{Logical. Should or not progress be printed.}
}
\value{
A data frame containing possible typographic errors,
taxonomic ranks extracted from 'info' and the phyletic
nature of each of them.
}
\description{
Function to account for the PUT status of the species in 'info',
spelling errors, putative MDCCs and the phyletic nature of
groups of PPCR species.
}
\examples{
cats.checked <- check_info(info=cats.info, tree=cats, sim=0.75)

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
