% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_to_branch.R
\name{custom.branch}
\alias{custom.branch}
\title{Function to add species at specified branches.}
\usage{
custom.branch(tree, edges, new.tip, rand.type = "random")
}
\arguments{
\item{edges}{matrix with 4 columns character vectors.
Columns 1 and 2 define the stem node of candidate branches and
columns 3 and 4 define the crown node
(mrca of species pairs). Note that a pair of nodes may not
necessarily define one single branch but a set of them}
}
\description{
Function to add species at specified branches.
}
