% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_to_branch.R
\name{custom.branch_col}
\alias{custom.branch_col}
\title{Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function}
\usage{
custom.branch_col(
  tree,
  edges,
  rand.type = "random",
  permitted.col = "red",
  forbidden.col = "black"
)
}
\description{
Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function
}
\examples{
set.seed(1)
}
