% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.info.R
\name{edit.info}
\alias{edit.info}
\title{Edit an 'info' file using a simple function.}
\usage{
\method{edit}{info}(info, PUTs, column = NULL, edit = NULL, remove.rows = FALSE)
}
\arguments{
\item{info}{An 'info' object.}

\item{PUTs}{Character vector with the PUT (or PUTs) to be edited or removed.}

\item{column}{Which column is to be edited for the specified PUTs. NULL value is only accepted if \code{remove.rows} is set to TRUE.}

\item{edit}{Any allowed value for the column of info that is to be edited. NULL value is only accepted if \code{remove.rows} is set to TRUE.}

\item{remove.rows}{If TRUE, the specified PUTs will be eliminated from the 'info' object. Default is FALSE.}
}
\value{
An 'info' object including the editions or deletions asked.
}
\description{
Function to edit randtip's 'info' file.
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
