% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_input.R
\name{edit.tree}
\alias{edit.tree}
\title{Edit backbone tree using a simple function.}
\usage{
\method{edit}{tree}(tree, tips, edit = NULL, remove.tips = FALSE)
}
\arguments{
\item{tree}{A backbone tree.}

\item{tips}{Character vector with the phylogenetic tips to be edited or removed.}

\item{edit}{A vector of the same length as \code{tips}, including the
new labels for the specified phylogenetic tips. NULL value is
only accepted if \code{remove.tips} is set to TRUE.}

\item{remove.tips}{If TRUE, the specified tips will be pruned
from the backbone tree. Default is FALSE.}
}
\value{
A backbone tree including the requested tip editions or deletions.
}
\description{
Function to edit a backbone tree.
}
\examples{
cats <- edit.tree(cats, tips="Felis_silvestris", 
edit= "Felis_silvestris_ssp._silvestris")

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
