% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_input.R
\name{edit_info}
\alias{edit_info}
\title{Function to edit info data frames.}
\usage{
edit_info(info, taxa, column = NULL, edit = NULL, remove.rows = FALSE)
}
\arguments{
\item{info}{An info data frame.}

\item{taxa}{Character vector with the taxa to be edited or removed.}

\item{column}{The name of the column to be edited for the specified taxa.
NULL is only accepted if \code{remove.rows} is TRUE.}

\item{edit}{Any allowed value for the column to be edited. NULL is only
accepted if \code{remove.rows} is TRUE.}

\item{remove.rows}{If TRUE, the specified taxa will be removed from the
info data frame (default is FALSE).}
}
\value{
An edited info data frame
}
\description{
Auxiliar function for smoothly editing the info data frame
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.info <- edit_info(cats.info, taxa= "Puma_concolor",
column = "subfamily", edit = "Felinae")
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
