% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish.info}
\alias{fish.info}
\title{fish.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 31521 species of vascular fish
included in 'fish' object. This object was created after The Fish Tree of Life taxonomy.

}
}
\usage{
fish.info
}
\description{
randtip 'info' file created after fish phylogeny
}
\keyword{datasets}
