% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{fish}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 31521 species of fish
Data obtained from https://github.com/megatrees/fish_20221117
}
}
\usage{
fish
}
\description{
Phylogenetic tree obtained from Rabosky et al. (2018).
}
\keyword{datasets}
