% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_to_branch.R
\name{get.candidate.edges}
\alias{get.candidate.edges}
\title{Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function}
\usage{
get.candidate.edges(tree, edges)
}
\description{
Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function
}
\examples{
set.seed(1)
tr <- ape::rcoal(25)
edges <- rbind(c("t22", "t14", "t22", "t22"),
               c("t7", "t10", "t11", "t3"))
candidate.edges <- get.candidate.edges(tr, edges)
# Now plot tree highlighting the candidate edges
edge_col <- ifelse(1:nrow(tr$edge) \%in\% candidate.edges, "red", "black")
edge_wid <- ifelse(1:nrow(tr$edge) \%in\% candidate.edges, 2, 1)     
plot(tr, edge.color = edge_col, edge.width = edge_wid)
# Using the same tree object tr we can select multiple candidate branches
edges <- rbind(c("t22", "t14", "t1", "t11"))
candidate.edges <- get.candidate.edges(tr, edges)
# Now plot tree highlighting the candidate edges
edge_col <- ifelse(1:nrow(tr$edge) \%in\% candidate.edges, "red", "black")
edge_wid <- ifelse(1:nrow(tr$edge) \%in\% candidate.edges, 2, 1)     
plot(tr, edge.color = edge_col, edge.width = edge_wid)
}
