% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get.clade}
\alias{get.clade}
\title{Function to extract a subtree representing a specified clade.}
\usage{
get.clade(info, tree, clade)
}
\arguments{
\item{info}{An 'info' (or 'input') object.}

\item{tree}{The original backbone tree to be split.}

\item{clade}{Clade to be extracted}
}
\value{
A list of four objects which will be used for automatic plotting using \code{\link{plot.clade}} function.
'Tree' will contain the splitted tree; 'info' will contain the handed info file; 'rank' will contain the taxonomic rank of the specified clade, and 'clade' will contain the clade name.
}
\description{
Obtain a phylogenetic tree representing a specified clade split from an original backbone tree.
Note that the tree splitting will be performed at the MRCA node of all the taxa included in the clade in the 'info' object.
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
