% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_clade}
\alias{get_clade}
\title{Function to extract clades from the backbone phylogeny.}
\usage{
get_clade(info, tree, group)
}
\arguments{
\item{info}{An 'info' (or 'input') data frame.}

\item{tree}{The backbone tree to be subsetted.}

\item{group}{Taxonomic group of species that defines the clade to extract.}
}
\value{
A list with four objects that will be used for automatic plotting
with the \code{\link{plot_clade}} function.
'Tree' is the subtree; 'info' is the handed info
data frame; 'rank' is the taxonomic rank of the specified group,
and 'clade' is the name of the clade. ## SAME AS ABOVE, SHOULD BE 'group'
}
\description{
This function serves to extract the clade in the backbone phylogeny that includes all the
species in the specified group (as defined in the info data frame).
}
\examples{
catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

felinae.clade <- get_clade(info=cats.info,
tree=cats, group="Felinae")

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
