% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{info2input}
\alias{info2input}
\title{Convert 'info' to 'input'.}
\usage{
info2input(info, tree, verbose = T)
}
\arguments{
\item{info}{An 'info' data frame, including all the customized binding
parameters.}

\item{tree}{Backbone tree.}
}
\value{
An 'input' data frame which can be fed to \code{rand_tip} function
alongside with a backbone tree to expand a tree.
}
\description{
Convert an 'info' object into an 'input' one.
}
\examples{
catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.input <- info2input(info=cats.info, tree=cats)

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
