% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.node}
\alias{is.node}
\title{this function reveals if a given node at a given tree is an internal node}
\usage{
is.node(tree, node)
}
\description{
this function reveals if a given node at a given tree is an internal node
}
