% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.info}
\alias{mammals.info}
\title{mammals.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 5911 species of mammals
included in 'mammals' object. Taxonomic information was retrieved from 'ncbi' database taxonomy.

}
}
\usage{
mammals.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the mammal phylogeny (Upham et al. 2019)
}
\keyword{datasets}
