% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{mammals}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 4099 species of mammals.
Data obtained from https://doi.org/10.1371/journal.pbio.3000494
}
}
\usage{
mammals
}
\description{
Phylogenetic tree obtained from Upham et al. (2019).
}
\keyword{datasets}
