% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mythology}
\alias{mythology}
\title{mythology}
\format{
A list with 5 slots

\describe{
\item{back.tree}{Hypothetical backbone tree}
\item{sp.list}{List of species to expanded the backbone tree}
\item{info.list}{Hypothetical 'info' data frame for use under "list" mode}
\item{info.backbone}{Hypothetical 'info' data frame for use under "backbone" mode}
\item{edges}{'edges' data frame to bind the PUT "Monoceros_x_alaricornus" using a customized set of edges}
}
}
\usage{
mythology
}
\description{
Data used in randtip tutorial (see Supplementary Material in https://doi.org/10.1101/2022.01.04.474924).
}
\keyword{datasets}
