% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyleticity.R
\name{phyleticity}
\alias{phyleticity}
\title{function phyleticity retrieves genus type given a certain tree}
\usage{
phyleticity(tree, genus)
}
\description{
function phyleticity retrieves genus type given a certain tree
}
\examples{
#phyleticity(Tree, "Invent")
}
