% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_table.R
\name{phylo.table}
\alias{phylo.table}
\title{function phylo.table returns a table which the main function can work
with (editable by hand!).}
\usage{
phylo.table(
  species,
  tree = NULL,
  taxonomy = NULL,
  genera.phyleticity = NULL,
  sep = "_"
)
}
\description{
function phylo.table returns a table which the main function can work
with (editable by hand!).
}
\examples{
#df <- phylo.table(species = checklist$match.name, Tree,  
#                  genera.phyleticity = tipos_generos, taxonomy = taxonomy)
}
