% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_taxonomy.R
\name{phylo.taxonomy}
\alias{phylo.taxonomy}
\title{function phylo.taxonomy returns a table with family, order and
class for every
genus given in the species vector.}
\usage{
phylo.taxonomy(species, db = "ncbi")
}
\description{
function phylo.taxonomy returns a table with family, order and
class for every
genus given in the species vector.
}
\examples{
phylo.taxonomy(species=c("Abies_pinsapo", "Pinus_pinaster", 
                         "Achyllea_milleifolium"))
}
