% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants.info}
\alias{plants.info}
\title{plants.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 73013 species of vascular plants
included in 'plants' object. Taxonomic information was retrieved from 'ncbi' database taxonomy.

}
}
\usage{
plants.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the plant phylogeny (Jin & Quian 2022)
}
\keyword{datasets}
