% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants}
\alias{plants}
\title{plants}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 74531 species of vascular plants.
Data obtained from https://github.com/jinyizju/V.PhyloMaker/tree/master/data
}
}
\usage{
plants
}
\description{
Phylogenetic tree obtained from Jin & Quian (2019), based on Smith & Brown (2018) and Zanne et al. (2014).
}
\keyword{datasets}
