% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_tree.R
\name{plot.clade}
\alias{plot.clade}
\title{Function to plot subtree}
\usage{
\method{plot}{clade}(
  get.clade.out,
  ppcr.col = "#4a8a21",
  nonppcr.col = "#48bce0",
  unknown.col = "#adadad",
  ...
)
}
\description{
Function to plot subtree
}
