% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_to_branch.R
\name{plot.custom.branch}
\alias{plot.custom.branch}
\title{Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function}
\usage{
\method{plot}{custom.branch}(
  tree,
  edges,
  rand.type = "random",
  permitted.col = "#03C03C",
  forbidden.col = "black",
  permitted.lwd = 1,
  forbidden.lwd = 1,
  ...
)
}
\description{
Function to help users visualize the candidate branches for tip insertion with
\code{\link{stick.to.branch}} function
}
\examples{
set.seed(1)
}
