% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clade}
\alias{plot_clade}
\title{Function to plot a subtree representing a specified clade.}
\usage{
plot_clade(
  get.clade.out,
  ppcr.col = "#4a8a21",
  nonppcr.col = "#48bce0",
  unknown.col = "#adadad",
  ...
)
}
\arguments{
\item{get.clade.out}{Output from \code{\link{get_clade}} function.}

\item{ppcr.col}{Color to represent tips included in the specified clade.
Default value is green.}

\item{nonppcr.col}{Color to represent tips included in a different clade
from the specified one (at the same taxonomic rank).
Default value is blue.}

\item{unknown.col}{Color to represent tips without taxonomic information
at the specifed clade's taxonomic rank.
Default value is grey}

\item{...}{Arguments to pass through \code{\link{plot.phylo}} function.}
}
\value{
A plot representing the clade specified in \code{\link{get_clade}}
function using the selected colors.
}
\description{
Plot a phylogenetic tree splitted from the backbone tree
using \code{get_clade} function.
}
\examples{
#First the clade information must be obtained

catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

felinae.clade <- get_clade(info=cats.info,
tree=cats, clade="Felinae")

#Then it can be plotted
plot_clade(felinae.clade, ppcr.col="green",
nonppcr.col="red",unknown.col="grey" )

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
