% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_branch.R
\name{plot_custom_branch}
\alias{plot_custom_branch}
\title{plot.custom.branch}
\usage{
plot_custom_branch(
  tree,
  edges,
  PUT = NULL,
  candidate.col = "#bf2828",
  forbidden.col = "#3d3d3d",
  candidate.lwd = 2,
  forbidden.lwd = 1,
  ...
)
}
\arguments{
\item{tree}{backbone tree of class phylo.}

\item{edges}{matrix with five columns (character vectors). Each row defines
a set of consecutive edges between two nodes, and multiple rows
can be used for the same PUT.
Column 1 includes the PUTs.
Columns 2 and 3 include a pair of species whose MRCA defines the older node,
whereas the MRCA of the species in columns 4 and 5 defines the younger node.
Using the same species in columns 4 and 5 will define a terminal node as the
younger node. If columns 2-3 are filled with a species and columns 4-5 are
filled with another species, then all edges below the MRCA of the two species
will be set as candidates. If the four columns are filled with the same
species, the PUT will be bound as sister to this species.}

\item{PUT}{If the \code{edges} data frame includes binding information for multiple
PUTs, specifies the PUT whose candidate edges are to be depicted.}

\item{candidate.col}{Color for candidate edges (default is "red").}

\item{forbidden.col}{Color for non-candidate edges (default is "black").}

\item{candidate.lwd}{Line width for candidate edges (default is 2).}

\item{forbidden.lwd}{ine width for non-candidate edges (default is 1).}

\item{...}{further arguments to be passed to \code{\link[ape]{plot.phylo}}.}
}
\description{
Auxiliar function to visualize candidate edges for PUT binding with \code{\link{custom_branch}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cats.edges <- data.frame(
 "PUT"= "Felis_catus",
 "parent1"= "Felis_silvestris",
 "parent2"= "Felis_silvestris",
 "child1"= "Felis_silvestris",
 "child2"= "Felis_silvestris")

#Depict candidate edges on the phylogeny
plot_custom_branch(tree=cats, edges=cats.edges)
\dontshow{\}) # examplesIf}
}
