% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_tree.R
\name{put.tip.col}
\alias{put.tip.col}
\title{Get PUT or placed color pattern}
\usage{
put.tip.col(newtree, oldtree, placed.col = "#adadad", put.col = "#C23B23")
}
\arguments{
\item{newtree}{An expanded phylogenetic tree.}

\item{oldtree}{The original backbone tree where the PUTs have been bound.}

\item{placed.col}{Color to plot phylogenetic tips which were already placed in the original backbone tree. Default value is grey.}

\item{put.col}{Color to plot bound PUTs in \code{new tree}. Default value is red.}
}
\value{
A vector of length equal to the number of tips in newtree, to be used after \'tip.color\' in plot.phylo function.
}
\description{
Get a color vector to use in \'tip.color\' argument within plot.phylo function
}
\author{
Ignacio Ramos-Gutiérrez, Rafael Molina-Venegas, Herlander Lima
}
