% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{put_tip_col}
\alias{put_tip_col}
\title{Get PUT or placed color pattern}
\usage{
put_tip_col(newtree, oldtree, placed.col = "#adadad", put.col = "#C23B23")
}
\arguments{
\item{newtree}{An expanded phylogenetic tree.}

\item{oldtree}{The original backbone tree where the PUTs have been bound.}

\item{placed.col}{Color to plot phylogenetic tips which were already placed in the original backbone tree.
Default value is grey.}

\item{put.col}{Color to plot bound PUTs in \code{new tree}. Default value is red.}
}
\value{
A vector of length equal to the number of tips in newtree, to be used after \code{tip.color}
in plot.phylo function.
}
\description{
Set color pattern to distinguish between phylogenetically placed taxa and PUTs when plotting trees
with the \code{plot.phylo} function of ‘ape’ R package using the \code{tip.color} argument
}
\examples{
#Perform a tree expansion
catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.input <- info2input(info=cats.info, tree=cats)

expanded.cats <- rand_tip(input=cats.input,
 tree=cats, rand.type = "polytomy",
 forceultrametric = TRUE)

#Set the colours for original tips and bound PUTs
cats.tip.cols <- put_tip_col(newtree = expanded.cats,
 oldtree = cats, placed.col="black", put.col="red")

#Plot the resulting tree visualizing original tips and PUTs
plot.phylo(expanded.cats, tip.color = cats.tip.cols)

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
