% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.tip.R
\name{rand.tip.multiple}
\alias{rand.tip.multiple}
\title{Expand multiple phylogenies binding PUTs to a backbone tree.}
\usage{
rand.tip.multiple(
  input,
  tree,
  number = 1,
  path = NULL,
  file = "randtip",
  rand.type = "random",
  polyphyly.scheme = "largest",
  use.paraphyletic = TRUE,
  use.singleton = TRUE,
  use.stem = FALSE,
  respect.mono = TRUE,
  respect.para = TRUE,
  clump.puts = TRUE,
  prob = TRUE,
  prune = TRUE,
  forceultrametric = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{An 'input' data frame obtained with \code{\link{info2input}} function.}

\item{tree}{A backbone tree.}

\item{number}{Integer. Number of phylogenies to be returned.}

\item{path}{Directory where to save the resulting trees. If path is set to NULL, they will not be automatically saved.}

\item{file}{File name for the resulting trees to be saved inside a folder with the same name.}

\item{rand.type}{For all PUTs not specified individually in 'input', which randomization type ("random" or
"polytomy") must be carried out. Default value is "random".}

\item{polyphyly.scheme}{For all PUTs not specified individually in 'input', which polyphyly
scheme ("largest", "complete" or "frequentist") must be used. Default value is "largest".}

\item{use.paraphyletic}{For all PUTs not specified individually in 'input', whether or not should paraphyletic
clades be taken into account or not. Default value is TRUE.}

\item{use.singleton}{For all PUTs not specified individually in 'input', should or not singleton MDCCs be
considered for binding as a sister species, or contrarily binding should be performed
anywhere below the parent node. Default value is TRUE.}

\item{use.stem}{For all PUTs not specified individually in 'input', whether or not should the stem branch be
considered as candidate for binding.  Default value is FALSE.}

\item{respect.mono}{For all PUTs not specified individually in 'input', whether or not monophyletic groups
should be respected when binding a PUT. Default value is TRUE.}

\item{respect.para}{For all PUTs not specified individually in 'input', whether or not paraphyletic groups
should be respected when binding a PUT. Default value is TRUE.}

\item{clump.puts}{For all PUTs not specified individually in 'input', whether or not co-ranked PUTs should be
clumped together in the phylogeny in case their taxonomic group is missing in the tree.
Will also clump conspecific PUTs. Default value is TRUE.}

\item{prob}{For all PUTs not specified individually in 'input', whether or not branch selection probability
must be proportional to branch length or equiprobable. Default value is TRUE.}

\item{prune}{Whether or not the newly expanded tree will include only the species in the user's list.
Default value is TRUE.}

\item{forceultrametric}{Whether or not the backbone tree will be forced to be ultrametric, only in case it is
not. Default value is FALSE.}

\item{verbose}{Whether or not to print information about the flow of the function. Default value is TRUE.}
}
\value{
A list containing a phylogeny in each element. In case 'path' is not set to NULL,
a folder will be created and tree files will be saved.
}
\description{
Expand multiple phylogenies binding PUTs to a backbone tree.
}
\examples{
expanded.cats.multiple <- rand.tip.multiple(input=cats.input, 
tree=cats, number = 10, path=getwd(), file="randtip",
rand.type = "random", forceultrametric = T)

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
