% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtip.R
\name{rand.tip}
\alias{rand.tip}
\title{randtip RANDOMIZATION FUNCTIONS}
\usage{
rand.tip(
  input,
  tree,
  rand.type = "random",
  polyphyly.scheme = "largest",
  use.paraphyletic = TRUE,
  use.singleton = TRUE,
  use.stem = FALSE,
  respect.mono = TRUE,
  respect.para = TRUE,
  clump.puts = TRUE,
  prob = TRUE,
  prune = TRUE,
  forceultrametric = FALSE,
  verbose = TRUE
)
}
\description{
randtip RANDOMIZATION FUNCTIONS
}
