% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.tip.R
\name{rand_tip}
\alias{rand_tip}
\title{Tree expansion function.}
\usage{
rand_tip(
  input,
  tree,
  rand.type = "random",
  polyphyly.scheme = "largest",
  use.paraphyletic = TRUE,
  use.singleton = TRUE,
  use.stem = FALSE,
  respect.mono = TRUE,
  respect.para = TRUE,
  clump.puts = TRUE,
  prob = TRUE,
  prune = TRUE,
  forceultrametric = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{An 'input' data frame obtained with \code{\link{info2input}} function.}

\item{tree}{A backbone tree.}

\item{rand.type}{For all PUTs not specified individually in 'input', which randomization type ("random" or
"polytomy") must be carried out. Default value is "random".}

\item{polyphyly.scheme}{For all PUTs not specified individually in 'input', which polyphyly
scheme ("largest", "complete" or "frequentist") must be used. Default value is "largest".}

\item{use.paraphyletic}{For all PUTs not specified individually in 'input', whether or not should paraphyletic
clades be taken into account or not. Default value is TRUE.}

\item{use.singleton}{For all PUTs not specified individually in 'input', should or not singleton MDCCs be
considered for binding as a sister species, or contrarily binding should be performed
anywhere below the parent node. Default value is TRUE.}

\item{use.stem}{For all PUTs not specified individually in 'input', whether or not should the stem branch be
considered as candidate for binding.  Default value is FALSE.}

\item{respect.mono}{For all PUTs not specified individually in 'input', whether or not monophyletic groups
should be respected when binding a PUT. Default value is TRUE.}

\item{respect.para}{For all PUTs not specified individually in 'input', whether or not paraphyletic groups
should be respected when binding a PUT. Default value is TRUE.}

\item{clump.puts}{For all PUTs not specified individually in 'input', whether or not co-ranked PUTs should be
clumped together in the phylogeny in case their taxonomic group is missing in the tree.
Will also clump conspecific PUTs. Default value is TRUE.}

\item{prob}{For all PUTs not specified individually in 'input', whether or not branch selection probability
must be proportional to branch length or equiprobable. Default value is TRUE.}

\item{prune}{Whether or not the newly expanded tree will include only the species in the user's list.
Default value is TRUE.}

\item{forceultrametric}{Whether or not the backbone tree will be forced to be ultrametric, only in case it is
not. Default value is FALSE.}

\item{verbose}{Whether or not to print information about the flow of the function. Default value is TRUE.}
}
\value{
An expanded phylogeny.
}
\description{
Expand a phylogeny binding PUTs to a backbone tree.
}
\details{
The optimum binding procedure might be different in each case,
given the phyletic nature of the MDCC to which the PUT should be bound to,
and different reasons might drive the user to choose one or another.
\itemize{
\item{polyphyly.scheme} {The resulting outputs may be very different depending on the
selected parameter. If 'complete' procedure is selected, the binding
will be performed as a monophyletic group using the MRCA of
all the known PPCR species. The 'largest' procedure, otherwise,
will use as binding group the largest monophyletic group formed
by PPCR species. Ultimately, the 'frequentist' option will
split the group in monophyletic and singleton chunks and
for every PUT to be bound to the MDCC, one of the chunks will
be selected weighting the probability by the number of original tips.}

\item{use.paraphyletic}{ This parameter should be used as TRUE in case we are certain of
the paraphyletic nature of one group, but disregarded if the
"intruder" group within the otherwise monophyletic group is possibly
due to an incorrect placement in the creation of the phylogeny.}

\item{forceultrametric} {It is important to note that forcing phylogenies to be ultrametric
in this way should not be taken as a formal statistical approach for inferring an
ultrametric tree but a method to be deployed whenever a genuinely ultrametric
phylogeny read from file fails due to issues related to numerical precision
(Revell, 2012). Thus, we strongly recommend the user to visually explore
phylogenetic trees that fail the ultrametricity test of check.info before
assuming the failure is due to numerical precision of computer machinery.}
}
}
\examples{

 catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.input <- info2input(info=cats.info, tree=cats)

expanded.cats <- rand_tip(input=cats.input,
 tree=cats, rand.type = "random",
  forceultrametric = FALSE)

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
