% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtip.R
\name{randtip}
\alias{randtip}
\title{randtip "MOTHER FUNCTION"}
\usage{
randtip(
  tree,
  species.table,
  type = c("random", "genus.polytomy", "family.polytomy", "order.polytomy",
    "class.polytomy"),
  aggregate.subspecies = TRUE,
  insertion = c("random", "middle", "long"),
  prob = TRUE,
  verbose = FALSE
)
}
\description{
randtip "MOTHER FUNCTION"
}
