% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamates.info}
\alias{squamates.info}
\title{squamates.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 4161 species of squamates
included in 'squamates' object. Taxonomic information was retrieved from 'ncbi' database taxonomy.

}
}
\usage{
squamates.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the squamate phylogeny (Pyron et al. 2013)
}
\keyword{datasets}
