% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamates}
\alias{squamates}
\title{squamates}
\format{
A phylo object
\describe{
Phylogenetic tree comprising 4161 species of squamates.
Data obtained from https://doi.org/10.1186/1471-2148-13-93
}
}
\usage{
squamates
}
\description{
Phylogenetic tree obtained from Pyron et al. (2013).
}
\keyword{datasets}
