% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_to_branch.R
\name{stick.to.branch}
\alias{stick.to.branch}
\title{Function to add species at specified branches.}
\usage{
stick.to.branch(tree, edges, new.tip = new.tip, prob = TRUE)
}
\arguments{
\item{edges}{matrix with 4 columns character vectors.
Columns 1 and 2 define the parent node of candidate branches and
columns 3 and 4 define the corresponding child node
(mrca of species pairs). Note that a pair of nodes may not
necessarily define one single branch but a set of them}
}
\description{
Function to add species at specified branches.
}
