% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_paraphyletic.R
\name{add.to.paraphyletic}
\alias{add.to.paraphyletic}
\title{Same process than paraphyletic node, bt detecting automatically the deep and shallow nodes}
\usage{
add.to.paraphyletic(
  tree,
  new.tip,
  prob = TRUE,
  intern.node = NULL,
  group.node = NULL
)
}
\arguments{
\item{intern.node}{node of the clade to insert the new tip. It must be used in conjuction
with group node}

\item{group.node}{mrca node of the paraphyletic clade. If no intern.node is supplied, a random
internal node is selected}
}
\description{
Same process than paraphyletic node, bt detecting automatically the deep and shallow nodes
}
\examples{
#add.to.paraphyletic(Tree, "Achillea_3_inventatus")
}
