% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_polyphyletic.R
\name{add.to.polyphyletic}
\alias{add.to.polyphyletic}
\title{A function to stick species at random within a polyphyletic clade}
\usage{
add.to.polyphyletic(
  tree,
  species,
  polyphyletic.insertion = c("freq", "large", "all")
)
}
\description{
A function to stick species at random within a polyphyletic clade
}
