% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibians.info}
\alias{amphibians.info}
\title{amphibians.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 7238 species of amphibians
included in 'amphibians' object. This object was created after 'ncbi' database taxonomy.

}
}
\usage{
amphibians.info
}
\description{
randtip 'info' file created after amphibians phylogeny
}
\keyword{datasets}
