% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds.info}
\alias{birds.info}
\title{birds.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 9993 species of birds included in
'birds' object. This object was mainly created after 'ncbi' database taxonomy.

}
}
\usage{
birds.info
}
\description{
randtip 'info' file created after birds phylogeny
}
\keyword{datasets}
