% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{birds}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 9993 species of birds.
Data obtained from https://github.com/megatrees/bird_20221117
}
}
\usage{
birds
}
\description{
Phylogenetic tree based in Jetz et al. (2012).
}
\keyword{datasets}
