% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_input.R
\name{build_info}
\alias{build_info}
\title{Create a 'info' data frame.}
\usage{
my.info <- build_info(species.list, tree, db="gbif", mode="list",
                   find.ranks = TRUE, interactive =FALSE,
                   genus = FALSE, prior.info = NULL, verbose = TRUE)
}
\arguments{
\item{species}{A character vector or a single-column data frame including
the species of interest. Word breakers must be blanks (" ")
or underscores ("_").}

\item{tree}{A 'phylo' object backbone tree. It can be set to NULL if
\code{mode} is set to "list".}

\item{find.ranks}{Logical. If TRUE, taxonomic information will be retrieved to
identify supra-generic MDCCs for the PUTs.}

\item{db}{Taxonomic data base to search into if \code{find.ranks} is
set to TRUE. Accepted values are 'ncbi' (default),
'itis', 'gbif' and 'bold'.}

\item{mode}{If mode is set to "list", the info file will be created
using only the species given in the \code{species} argument.
If "backbone" mode is specified, 'info' will also include all
the tips included in the backbone tree.}

\item{interactive}{Logical. Whether or not ambiguous species names will
be resolved manually by the user or filled in
automatically with 'NA' when retrieving taxonomic
information.}

\item{genus}{Logical. Whether or not a genus-level backbone tree is to
be expanded. If set to TRUE, all tips in the backbone tree
and taxa in the species vector must represent genera.}

\item{prior.info}{A previously created 'info' file used to fill blanks
previously to the taxonomic database search.}

\item{verbose}{Logical. Should or not progress be printed.}
}
\value{
A randtip 'info' data frame
}
\description{
Function to create an 'info' object given a list of species.
}
\examples{
# Create a list of species to include in the resulting tree
 catspecies <- c("Lynx_lynx",
"Panthera_uncia",
"Panthera_onca",
"Felis_catus",
"Puma_concolor",
"Lynx_canadensis",
"Panthera_tigris",
"Panthera_leo",
"Felis_silvestris")

#Create the 'info' file
cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
