% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cats}
\alias{cats}
\title{Phylogeny of felids}
\format{
A phylogenetic tree with 14 tips and 13 internal nodes
\describe{
Phylogenetic tree with 14 species of felids (Felidae).
This is a subtree extracted from the complete mammal phylogeny by Upham et al. (2019)
}
}
\usage{
cats
}
\description{
Example phylogeny for randtip package
}
\keyword{datasets}
