% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{check.info}
\alias{check.info}
\title{Function to check randtip's files.}
\usage{
check.info(info, tree, sim = 0.8)
}
\arguments{
\item{info}{An 'info' object.}

\item{tree}{The original backbone tree.}

\item{sim}{Name similarity threshold to detect possible misspellings
on tip labels. Default value is 0.8.
See \link[stringdist]{stringsim} for details.}
}
\value{
A data frame containing possible typographic errors, taxonomic ranks extracted from 'info' and the phyletic nature of each of them.
#'
}
\description{
Function to account for the PUT status of the species in 'info',
spelling errors, putative MDCCs and the phyletic nature of
groups of PPCR species.
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
