% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.branch.R
\name{custom.branch}
\alias{custom.branch}
\title{custom.branch}
\usage{
custom.branch(
  tree,
  edges,
  rand.type = "random",
  forceultrametric = F,
  prob = T
)
}
\arguments{
\item{tree}{"phylo" object used as backbone tree.}

\item{edges}{matrix with 5 character vector columns.
Column 1 secifies the PUT to which the candidate branches in the row refer to.
Columns 2 and 3 must represent two species whose MRCA represents the stem node of candidate branches and
columns 4 and 5 two species whose MRCA define the crown node.
Note that a pair of nodes may not necessarily define one single branch but a set of them.
Two identical species will define a tree tip rather than an internal node.
If the pairs of species in columns 2&3 and 4&5 are the same, the set of branches to select will be defined as the complete clade below ther MRCA.}

\item{rand.type}{"random" or "polytomy". Default value is "random".}

\item{forceultrametric}{Whether or not the backbone tree will be forced to be ultrametric, only in case it is not. Default value is FALSE.}

\item{prob}{Whether or not branch selection probability must be proportional to branch length or equiprobable. Default value is TRUE.}
}
\value{
An expanded phylogeny.
}
\description{
Bind PUTs at completely customized tree branches
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
