% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_input.R
\name{edit_info}
\alias{edit_info}
\title{Edit an 'info' file using a simple function.}
\usage{
edit_info(info, taxa, column = NULL, edit = NULL, remove.rows = FALSE)
}
\arguments{
\item{info}{An 'info' object.}

\item{taxa}{Character vector with the taxa to be edited or removed.}

\item{column}{The column name to be edited for the specified taxa.
NULL value is only accepted if \code{remove.rows} is set to TRUE.}

\item{edit}{Any allowed value for the column of info that is to be edited.
NULL value is only accepted if \code{remove.rows} is set to TRUE.}

\item{remove.rows}{If TRUE, the specified taxa will be eliminated from the
'info' object. Default is FALSE.}
}
\value{
An 'info' object including the editions or deletions asked.
}
\description{
Function to edit randtip's 'info' file.
}
\examples{
 catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.info <- edit_info(cats.info, taxa= "Puma_concolor",
column = "subfamily", edit = "Felinae")

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
