% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish.info}
\alias{fish.info}
\title{fish.info}
\format{
An info data frame
\describe{
Data frame with taxonomic information on the 31521 species of fishes
included in 'fish' object. Taxonomic information was retrieved from The Fish Tree of Life website.

}
}
\usage{
fish.info
}
\description{
Info data frame with taxonomic information retrieved for the species in the fish phylogeny (Rabosky et al. 2018)
}
\keyword{datasets}
