% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{fish}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 11638 species of birds.
Data obtained from https://fishtreeoflife.org/downloads/
}
}
\usage{
fish
}
\description{
Phylogenetic tree obtained from The Fish Tree of Life.
}
\keyword{datasets}
