% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{info2input}
\alias{info2input}
\title{Convert 'info' to'input'.}
\usage{
info2input(info, tree)
}
\arguments{
\item{info}{An 'info' data frame, including all the customized binding parameters.}

\item{tree}{Backbone tree.}
}
\value{
An 'input' data frame which can be fed to \code{rand.tip} alongside with a backbone tree to expand a tree.
}
\description{
Convert an 'info' object into an 'input' one.
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
