% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.info}
\alias{mammals.info}
\title{mammals.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 4099 species of vascular mammals
included on 'mammals' object. This object was created after 'ncbi' database taxonomy.

}
}
\usage{
mammals.info
}
\description{
randtip 'info' file created after mammals phylogeny
}
\keyword{datasets}
