% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{mammals}
\format{
a Phylo object
\describe{
Phylogenetic tree comprising 5911 species of mammals.
Data obtained from https://github.com/megatrees/mammal_20221117
}
}
\usage{
mammals
}
\description{
Phylogenetic tree obtained from Upham et al. (2019).
}
\keyword{datasets}
