% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mythology}
\alias{mythology}
\title{mythology}
\format{
A list with 5 slots

\describe{
\item{back.tree}{Imaginary backbone tree where PUTs will be bound.}
\item{sp.list}{List of species to include in the final expanded tree}
\item{info.list}{'info' file resulting after running build.info using mode="list"}
\item{info.backbone}{'info' file resulting after running build.info using mode="backbone"}
\item{edges}{'edges' file used for manual selection of branches where to bind the PUT "Monoceros_x_alaricornus"}
}
}
\usage{
mythology
}
\description{
Data used in randtip tutorial (see Supplementary Material in https://doi.org/10.1101/2022.01.04.474924).
}
\keyword{datasets}
