% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants.info}
\alias{plants.info}
\title{plants.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 74531 species of vascular plants
included on 'plants' object. This object was created after 'ncbi' database taxonomy.

}
}
\usage{
plants.info
}
\description{
randtip 'info' file created after plants phylogeny
}
\keyword{datasets}
