% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants}
\alias{plants}
\title{plants}
\format{
A phylo object
\describe{
Phylogenetic tree comprising 73011 species of vascular plants.
Data obtained from https://github.com/megatrees/plant_20221214
}
}
\usage{
plants
}
\description{
Phylogenetic tree obtained from Jin & Quian (2022), based on Smith & Brown (2018) and Zanne et al. (2014).
Nomenclature stands for World Checklist of Vascular Plants (WCVP).
}
\keyword{datasets}
