% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_tree.R
\name{plot.clade}
\alias{plot.clade}
\title{Function to plot subtree}
\usage{
\method{plot}{clade}(
  get.clade.out,
  corankedtaxa.col = "#03C03C",
  intruder.col = "#C23B23",
  stowaway.col = "black",
  ...
)
}
\description{
Function to plot subtree
}
