% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clade}
\alias{plot_clade}
\title{Function to plot clades extracted from the backbone phylogeny.}
\usage{
plot_clade(
  get.clade.out,
  ppcr.col = "#4a8a21",
  nonppcr.col = "#48bce0",
  unknown.col = "#adadad",
  ...
)
}
\arguments{
\item{get.clade.out}{Output of the \code{\link{get_clade}} function (a phylogenetic clade)}

\item{ppcr.col}{Color for the tips representing species placed in the specified taxonomic group (default is green).}

\item{nonppcr.col}{Color for the tips representing other species but those placed in the specified taxonomic groupnt clade (default is blue).}

\item{unknown.col}{Color for the tips representing species without taxonomic information (default is grey).}

\item{...}{Further arguments to pass through \code{\link{plot.phylo}} function.}
}
\value{
Less inclusive clade that includes all the species in the specified taxonomic group.
}
\description{
Function to plot clades extracted from the backbone phylogeny.
}
\examples{
catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

felinae.clade <- get_clade(info=cats.info,
tree=cats, group="Felinae")

plot_clade(felinae.clade, ppcr.col="green",
nonppcr.col="red",unknown.col="grey" )

}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
