% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{put_tip_col}
\alias{put_tip_col}
\title{Get PUT or placed color pattern}
\usage{
put_tip_col(newtree, oldtree, placed.col = "#adadad", put.col = "#C23B23")
}
\arguments{
\item{newtree}{An expanded phylogenetic tree.}

\item{oldtree}{The original backbone tree.}

\item{placed.col}{Color for phylogenetic tips that were already placed in the backbone tree (default is grey).}

\item{put.col}{Color for phylogenetic tips representing PUTs in \code{new tree} (default is red).}
}
\value{
A vector of length equal to the number of tips in newtree that defines the argument \code{tip.color} in plot.phylo function.
}
\description{
Set a color pattern to distinguish between phylogenetically placed taxa and PUTs when plotting expanded trees
with the \code{plot.phylo} function of ‘ape’ R package
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

catspecies <- c("Lynx_lynx", "Panthera_uncia",
"Panthera_onca", "Felis_catus", "Puma_concolor",
"Lynx_canadensis", "Panthera_tigris", "Panthera_leo",
"Felis_silvestris")

cats.info <- build_info(species=catspecies, tree= cats,
     find.ranks=TRUE, db="ncbi", mode="backbone")

cats.input <- info2input(info=cats.info, tree=cats)

expanded.cats <- rand_tip(input=cats.input,
 tree=cats, rand.type = "polytomy",
 forceultrametric = TRUE)

#Set color pattern for original tips and PUTs, respectively
cats.tip.cols <- put_tip_col(newtree = expanded.cats,
 oldtree = cats, placed.col="black", put.col="red")

plot.phylo(expanded.cats, tip.color = cats.tip.cols)
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutierrez, Rafael Molina-Venegas, Herlander Lima
}
