% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtip.R
\name{rand.tip}
\alias{rand.tip}
\title{randtip RANDOMIZATION FUNCTIONS}
\usage{
rand.tip(
  DF1,
  tree,
  rand.type = "random",
  polyphyly.scheme = "large",
  use.paraphyletic = TRUE,
  use.singleton = TRUE,
  respect.mono = TRUE,
  respect.para = TRUE,
  clump.PUTs = TRUE,
  prune = TRUE,
  forceultrametric = TRUE,
  verbose = FALSE
)
}
\description{
randtip RANDOMIZATION FUNCTIONS
}
