% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamates.info}
\alias{squamates.info}
\title{squamates.info}
\format{
a 'info' data frame
\describe{
Data frame with taxonomic information on the 4161 species of vascular squamates
included on 'squamates' object. This object was created after 'ncbi' database taxonomy.

}
}
\usage{
squamates.info
}
\description{
randtip 'info' file created after squamates phylogeny
}
\keyword{datasets}
